fix: fix native dialogs being freed after lifetime ended
Native dialogs are not top-level widgets,
and thus we own it and GTK does not keep it alive.

This led to dialogs being freed after its Rust lifetime ended,
which makes the dialog never get to be shown to the user.

The fix doesn't include workspace.rs.
That part will be fixed in the workspace PR.

See https://docs.gtk.org/gtk4/class.FileChooserNative.html