fix: fix files not showing up in the file tree
Firstly, as the title suggests.
GtkTreeStore requires columns to be specified.

See:
https://www.mono-project.com/docs/gui/gtksharp/widgets/treeview-tutorial

Secondly, add more debug logs in the workspace code.
I marked the main thread code to be CLIENT and the async code to be SERVER.

Thirdly, change the tree view code to be in the UI file.

This patch doesn't make it works though.
The current issues are:
- The tree view makes the window end indefinitely until the content ends.
 I should implement a scrollbar for it.
- Folders aren't displayed, and files are attached to the root.
 When the parent iter can not be found,
 I made it so that the folders' files are attached to the root.
 We'll have to figure why we can't found the parent iter.