chore: make icon simplified and colorful
Based on feedbacks from some friends, make icon simplified and not busy
with six lines and not nine.

Additionally, add "syntax highlighting" feeling to the icon.
From feedbacks, I think it's good if I can change the code in the icon
to feel like an echidna and not something of "a dinosaur opening its
mouth" or "a bird".