feat: add activity bar buttons
Add two buttons at the bottom of the activity bar:
Profile and Settings buttons.
These are still at prototyping stage and isn't complete.

But the Profile button is already a usable MenuButton.

TODO: The button will need to open a popover.
The popover will need to display some user information
and a short button to open Profile settings.